Unify two `requirements.txt` files
Currently there are two `requirements.txt` files: one for the documentation and one for the code that lives in `.devcontainer`.

I think it makes sense to keep all the requirements together in a single file in the root folder. Otherwise users who aren't running the code in a container might miss that there's a second `requirements.txt` file lurking there.